requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
geoalchemy2==0.14.3
python-dotenv

# Development and testing
pytest==7.4.4
pytest-cov==4.1.0
httpx==0.26.0
ruff==0.1.14
